SER-551 more refactor
