+++ front/main/app/components/image-review.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import ImageReviewItemModel from '../models/image-review-item';
+import ImageReviewItemModel from '../models/image-review/image-review-item';
 
 export default Ember.Component.extend({
 	classNames: ['image-review'],
@@ -8,9 +8,9 @@ export default Ember.Component.extend({
 	actions: {
 		showModal(popupModel) {
 			ImageReviewItemModel.getImageContext(popupModel.imageId).then((data) => {
-				popupModel.data = data;
-				popupModel.isContextProvided = Boolean(data.context);
-				popupModel.isContextLink = new RegExp('(http|https)?:\/\/[^\s]+').test(data.context);
+				popupModel.set('data', data);
+				popupModel.set('isContextProvided', Boolean(data.context));
+				popupModel.set('isContextLink', new RegExp('(http|https)?:\/\/[^\s]+').test(data.context));
 			}).finally(() => {
 				this.setProperties({
 					thumbnailModel: popupModel,
