+++ front/main/app/models/image-review/image-review-item.js
